fiddle with root visibility

+++ widgets/cRootContainer.h
@@ -6,7 +6,7 @@ class cRootContainer : public cContainer {
 public:
   //{{{
   cRootContainer (uint16_t width, uint16_t height) :
-    cContainer(width / getBoxHeight(), height / getBoxHeight()) {}
+    cContainer(width / (float)getBoxHeight(), height / (float)getBoxHeight()) {}
   //}}}
   virtual ~cRootContainer() {}
